pyproject.toml
[tool.poetry]
name = "proyecto"
version = "0.1.0"
description = ""
authors = ["eporrasm <vals4oo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
pandas = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
